refactor: 리팩토링 데이 반영때 놓친 applicaton 패키지 분리
