+++ server/src/main/java/com/ahmadda/application/scheduler/EventNotificationScheduler.java
@@ -1,4 +1,4 @@
-package com.ahmadda.application;
+package com.ahmadda.application.scheduler;
 
 import com.ahmadda.domain.event.Event;
 import com.ahmadda.domain.event.EventRepository;
