+++ server/src/main/java/com/ahmadda/infra/notification/push/config/PushConfig.java
@@ -6,6 +6,7 @@ import com.ahmadda.infra.notification.push.FcmPushErrorHandler;
 import com.ahmadda.infra.notification.push.FcmPushNotifier;
 import com.ahmadda.infra.notification.push.FcmRegistrationTokenRepository;
 import com.ahmadda.infra.notification.push.MockPushNotifier;
+import jakarta.persistence.EntityManager;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
@@ -25,12 +26,14 @@ public class PushConfig {
     @ConditionalOnProperty(name = "push.mock", havingValue = "false", matchIfMissing = true)
     public PushNotifier fcmPushNotifier(
             final FcmRegistrationTokenRepository fcmRegistrationTokenRepository,
-            final NotificationProperties notificationProperties
+            final NotificationProperties notificationProperties,
+            final EntityManager em
     ) {
         return new FcmPushNotifier(
                 fcmRegistrationTokenRepository,
                 new FcmPushErrorHandler(fcmRegistrationTokenRepository),
-                notificationProperties
+                notificationProperties,
+                em
         );
     }
 }
